Func tests use fedora-heat-test-image
This runs the existing functional tests using the images built in
http://tarballs.openstack.org/heat-test-image/

This image should be a suitable alternative to
Fedora-x86_64-20-20140618-sda, with the extra benefit that it is
prepared to install the heat config agent projects during boot
so that test_server_software_config can be enabled.

This will also allow devstack to no longer load
Fedora-x86_64-20-20140618-sda onto the nodepool images which will reduce
gate resource consumption.

Change-Id: I6041b8d6e7e9422f6e220d7aef0ca38857085e4b