Remove capacity check from AWS::AG update
When heat is trying to update autoscaling group resource with
the same template an exception is raised because nested stack
is not in (UPDATE, COMPLETE) state. It happens because heat
doesn't change the state of nested stacks if size of Autoscaling
group has not been changed. So fix allows handle_update in
AWS::AutoscalingGroup to launch resize despite of changes
in AG capacity.

Change-Id: Ibb71f31f2ec7d1d37b995323a2474b799ea498bb
Closes-bug: #1429134