Wait for the stack lock to be released
Sometimes, when trying to update a stack in a *_COMPLETE/FAILED status,
it can fail since the stack lock is not released yet. As a short-term
workaround, I've put in place an ugly hack that will simply retry the
update if the operation failed because of the stack lock. The downside of
this is that we can't now add functional tests which will test updates on
Stack Lock itself since we now simply just ignore the HTTPConflict Exception
on updates.

I couldn't think of another solution that doesn't include a giant refactoring
that will solve this problem.

PS: Convergence doesn't have the notion of a Stack-Lock, so that should
take care of this by default.

Partially-Closes Bug: #1450314

Change-Id: Ib1a9d5c425285ebcffcb9ff8a362a56fd8f3574a