Allow nested validation for template-validate
Adds an optional flag that causes template-validate to recursively
validate nested stacks and expose a schema of the parameters for
all stacks in the tree.  This uses a similar method to stack preview
in that it builds the tree in memory so we can then walk it and
extract the parameter data we need.

Change-Id: I7e3d752124997a0a7b8ff53e774ed8034dd2f81d
Co-Authored-By: Jay Dobies <jdobies@redhat.com>
Implements-blueprint: nested-validation