Explicitly set heat.conf logging_exception_prefix
Gate logs for heat jobs really need to log stack traces with TRACE
instead of ERROR because some traces are being printed by passing
exc_info=True to LOG.info so they are not actual errors.

TRACE should be the default for oslo.log but something in the gate
setup is setting it to ERROR. Setting explicitly here restores the
desired pattern and protects against gate default overrides.

Change-Id: Id85287324fa6dc79d67fb1ed37c3f256d661133e
Closes-Bug: #1606404