Remove oslo-incubator use in other client plugins
oslo incubator code is planned to be removed from the clients
this cycle. Remove the redundant usage of it from other client
plugins.

This also includes an unrelated change(to remove the vpnaas
post configuration, as we have removed the plugin from the
gate jobs), to fix the gate.

Last but not least, it limits the number of connections to the database
to work around a convergence issue.

Change-Id: I01c46197dfebb1eb674817ce81792972a06e529f