Check for correct Neutron exceptions harder
The change to stop using the Nova passthrough API for networking in
e8e0a2483a8cd0edbd0ec77d92d638974c72f290 changed the type of exceptions
expected in a couple of places, without changing the exception handling
code accordingly.

7b17c8b4eea1091ff91a8a3d9e41a75256ee9302 changed the exception handling
in one case, but continued to raise the exception instead of suppressing
it.

This patch suppresses the exception, and makes the same change to
another case of the problem.

Use a different network name for test_create_server_with_port instead of
'my_net' which is also used by one of the other tests in the same file,
to avoid the conflict at its source.

Change-Id: I2f1210e546a5ad7eef3ffd6537335049120b2d2c
Co-Authored-By: Zane Bitter <zbitter@redhat.com>
Closes-Bug: #1694412