Add a functional test for nested get_attr functions
Nesting get_attr functions (i.e. using a get_attr in the arguments to
another get_attr) is unwise, and we probably should have prohibited it
somehow. Nonetheless, we didn't and some users (including but not
necessarily limited to TripleO) depend on this working. Since this would be
easy to break by accident, add a functional test to make sure we can only
break it on purpose.

Change-Id: I234a0762a00815af86b5711e7494366c16e717c2